// src/app/models/user.ts
export interface User {
    email? : string;
    user_id?: string;
    district?: string;
    district_id?: number;
    taluka?: string;
    taluka_id?: number;
    type?: string;
    village_city_id?: number;
    village_city?: string;
    municipal_gp_type?: string;
    municipal_corp?: string;
}